% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Define server logic required to draw a histogram}
\usage{
server(input, output)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}
}
\value{
Server function for a Shiny app.
}
\description{
Define server logic required to draw a histogram
}
